requirements.txt
websocket
finnhub-python
time
numpy==1.23
